extract core/proyecciones.py
